LayoutTests/platform/mac/imported/w3c/web-platform-tests/content-security-policy/connect-src/connect-src-eventsource-allowed.sub-expected.txt

CONSOLE MESSAGE: Refused to connect to http://www1.localhost:8800/content-security-policy/connect-src/resources/simple-event-stream because it does not appear in the connect-src directive of the Content Security Policy.

FAIL Expecting logs: ["allowed"] assert_unreached: unexpected log: blocked Reached unreachable code
